Revert "convert to typescript (#11)"
This reverts commit a85272f5a11b101e35fcb0fbe6575ebe54aee6d5.
