+++ docs/js/angle.js
@@ -1,24 +1,37 @@
 /**
- * Represents an angle
+ * @class Angle - represents an angle.
  */
 class Angle {
-    /**
-     * @param angle The angle, in degrees.
-     */
-    constructor(angle) {
-        this.angle = Math.abs(angle) % 360;
-    }
-    get degrees() {
-        return this.angle;
-    }
-    /**
-     * @returns The angle, in radians.
-     */
-    get radians() {
-        return this.angle * Math.PI / 180;
-    }
-    compare(angle) {
-        return this.degrees - angle.degrees;
-    }
+  /**
+   * @private
+   * @type {int} - The angle, in degrees.
+   */
+  #angle;
+
+  /**
+   * @param {number} angle - The angle, in degrees.
+   */
+  constructor(angle) {
+    this.#angle = Math.abs(angle) % 360;
+  };
+
+  /**
+   * @returns {number} - The angle, in degrees.
+   */
+  get degrees() {
+    return this.#angle;
+  };
+
+  /**
+   * @returns {number} - The angle, in radians.
+   */
+  get radians() {
+    return this.#angle * Math.PI / 180;
+  };
+
+  compare(angle) {
+    return this.degrees - angle.degrees;
+  };
 }
+
 export default Angle;
