+++ package.json
@@ -4,10 +4,9 @@
     "description": "Replacement for <https://wheelofnames.com>.",
     "main": "index.js",
     "scripts": {
-        "start": "concurrently npm:build:css vite npm:build:tsw",
-        "build:css": "sass --watch --no-source-map src/scss/main.scss docs/css/main.css",
-        "build:tsw": "tsc --watch",
-        "build:ts": "tsc"
+        "start": "concurrently npm:build:css vite",
+
+        "build:css": "sass --watch --no-source-map src/scss/main.scss docs/css/main.css"
     },
     "repository": {
         "url": "git@github.com:matthewdenaburgvc/wheel.git",
@@ -26,17 +25,13 @@
         "singleAttributePerLine": true
     },
     "devDependencies": {
-        "@types/canvas-confetti": "^1.9.0",
-        "@types/jquery": "^3.5.32",
         "concurrently": "^9.1.0",
         "prettier": "^3.3.3",
         "sass": "^1.81.0",
-        "typescript": "^5.8.2",
         "vite": "^5.4.11",
         "vite-plugin-sass": "^0.1.0"
     },
     "dependencies": {
-        "canvas-confetti": "^1.9.3",
         "jquery": "^3.7.1",
         "jquery-ui": "^1.13.3"
     }
